+++ cmd_queue/airflow_queue.py
@@ -74,7 +74,7 @@ class AirflowQueue(base_queue.Queue):
         >>> job1 = self.submit('echo hi 1 && true')
         >>> job2 = self.submit('echo hi 2 && true')
         >>> job3 = self.submit('echo hi 3 && true', depends=job1)
-        >>> self.rprint()
+        >>> self.print_commands()
         >>> self.write()
         >>> self.run()
         >>> #self.run()
@@ -227,7 +227,7 @@ class AirflowQueue(base_queue.Queue):
         self.named_jobs[job.name] = job
         return job
 
-    def rprint(self, with_status=False, with_gaurds=False, with_rich=1, colors=1):
+    def print_commands(self, with_status=False, with_gaurds=False, with_rich=None, colors=1, style='auto'):
         r"""
         Print info about the commands, optionally with rich
 
@@ -236,25 +236,32 @@ class AirflowQueue(base_queue.Queue):
             >>> from cmd_queue.airflow_queue import *  # NOQA
             >>> self = AirflowQueue()
             >>> self.submit('date')
-            >>> self.rprint()
+            >>> self.print_commands()
             >>> self.run()
         """
+        style = self._coerce_style(style, with_rich, colors)
+
         code = self.finalize_text()
-        if with_rich:
+
+        if style == 'rich':
             from rich.panel import Panel
             from rich.syntax import Syntax
             from rich.console import Console
             console = Console()
             console.print(Panel(Syntax(code, 'python'), title=str(self.fpath)))
             # console.print(Syntax(code, 'bash'))
-        else:
+        elif style == 'colors':
             header = f'# --- {str(self.fpath)}'
-            if colors:
-                print(ub.highlight_code(header, 'python'))
-                print(ub.highlight_code(code, 'python'))
-            else:
-                print(header)
-                print(code)
+            print(ub.highlight_code(header, 'python'))
+            print(ub.highlight_code(code, 'python'))
+        elif style == 'plain':
+            header = f'# --- {str(self.fpath)}'
+            print(header)
+            print(code)
+        else:
+            raise KeyError(f'Unknown style={style}')
+
+    rprint = print_commands  # backwards compat
 
 
 def demo():